+++ Trabajo Practico Integrador Programacion 2/src/Models/Paciente.java
@@ -9,16 +9,16 @@ package Models;
  * @author Usuario
  */
 public class Paciente {
-    private int idPaciente;
+    private Long idPaciente;
     private String nombre;
     private String apellido;
-    private int dni;
+    private Long dni;
     private String direccion;
     private java.sql.Date fechaNacimiento;
 
     public Paciente() {}
 
-    public Paciente(String nombre, String apellido, int dni, String direccion, java.sql.Date fechaNacimiento) {
+    public Paciente(String nombre, String apellido, Long dni, String direccion, java.sql.Date fechaNacimiento) {
         this.nombre = nombre;
         this.apellido = apellido;
         this.dni = dni;
@@ -27,8 +27,8 @@ public class Paciente {
     }
 
     // Getters y Setters
-    public int getIdPaciente() { return idPaciente; }
-    public void setIdPaciente(int idPaciente) { this.idPaciente = idPaciente; }
+    public Long getIdPaciente() { return idPaciente; }
+    public void setIdPaciente(Long idPaciente) { this.idPaciente = idPaciente; }
 
     public String getNombre() { return nombre; }
     public void setNombre(String nombre) { this.nombre = nombre; }
@@ -36,8 +36,8 @@ public class Paciente {
     public String getApellido() { return apellido; }
     public void setApellido(String apellido) { this.apellido = apellido; }
 
-    public int getDni() { return dni; }
-    public void setDni(int dni) { this.dni = dni; }
+    public Long getDni() { return dni; }
+    public void setDni(Long dni) { this.dni = dni; }
 
     public String getDireccion() { return direccion; }
     public void setDireccion(String direccion) { this.direccion = direccion; }
